Add files via upload
I am adding the code used to train the data and create an ai that can correctly pick what the person is trying to say. I also added the code for the actual chat bot once the training script is run the chat bot should also run successfully. The final file i am adding in this commit is the json file for the actual pattern recognition. so basically what the responses are going to be from the chat bot and what leads to them.